requirements.txt
streamlit
google-generativeai
python-dotenv
Pillow
pydantic
tenacity
pytest
pytest-mock
